TST/CLN: Move raising apply tests to test_invalid_arg (#40213)
