+++ pandas/tests/apply/test_series_apply.py
@@ -20,7 +20,6 @@ from pandas import (
     timedelta_range,
 )
 import pandas._testing as tm
-from pandas.core.base import SpecificationError
 
 
 def test_series_map_box_timedelta():
@@ -269,20 +268,6 @@ def test_demo():
     expected = Series([0], index=["foo"], name="series")
     tm.assert_series_equal(result, expected)
 
-    # nested renaming
-    msg = "nested renamer is not supported"
-    with pytest.raises(SpecificationError, match=msg):
-        s.agg({"foo": ["min", "max"]})
-
-
-def test_multiple_aggregators_with_dict_api():
-
-    s = Series(range(6), dtype="int64", name="series")
-    # nested renaming
-    msg = "nested renamer is not supported"
-    with pytest.raises(SpecificationError, match=msg):
-        s.agg({"foo": ["min", "max"], "bar": ["sum", "mean"]})
-
 
 def test_agg_apply_evaluate_lambdas_the_same(string_series):
     # test that we are evaluating row-by-row first
@@ -437,30 +422,6 @@ def test_agg_cython_table_transform(series, func, expected):
     # pandas.core.base.SelectionMixin._cython_table (cumprod, cumsum)
     result = series.agg(func)
     tm.assert_series_equal(result, expected)
-
-
-@pytest.mark.parametrize(
-    "series, func, expected",
-    chain(
-        tm.get_cython_table_params(
-            Series("a b c".split()),
-            [
-                ("mean", TypeError),  # mean raises TypeError
-                ("prod", TypeError),
-                ("std", TypeError),
-                ("var", TypeError),
-                ("median", TypeError),
-                ("cumprod", TypeError),
-            ],
-        )
-    ),
-)
-def test_agg_cython_table_raises(series, func, expected):
-    # GH21224
-    msg = r"[Cc]ould not convert|can't multiply sequence by non-int of type"
-    with pytest.raises(expected, match=msg):
-        # e.g. Series('a b'.split()).cumprod() will raise
-        series.agg(func)
 
 
 def test_series_apply_no_suffix_index():
@@ -732,9 +693,6 @@ def test_map_categorical():
     tm.assert_series_equal(result, exp)
     assert result.dtype == object
 
-    with pytest.raises(NotImplementedError, match=tm.EMPTY_STRING_PATTERN):
-        s.map(lambda x: x, na_action="ignore")
-
 
 def test_map_datetimetz():
     values = pd.date_range("2011-01-01", "2011-01-02", freq="H").tz_localize(
@@ -755,9 +713,6 @@ def test_map_datetimetz():
     result = s.map(lambda x: x.hour)
     exp = Series(list(range(24)) + [0], name="XX", dtype=np.int64)
     tm.assert_series_equal(result, exp)
-
-    with pytest.raises(NotImplementedError, match=tm.EMPTY_STRING_PATTERN):
-        s.map(lambda x: x, na_action="ignore")
 
     # not vectorized
     def f(x):
@@ -825,14 +780,6 @@ def test_map_float_to_string_precision():
     result = ser.map(lambda val: str(val)).to_dict()
     expected = {0: "0.3333333333333333"}
     assert result == expected
-
-
-def test_map_with_invalid_na_action_raises():
-    # https://github.com/pandas-dev/pandas/issues/32815
-    s = Series([1, 2, 3])
-    msg = "na_action must either be 'ignore' or None"
-    with pytest.raises(ValueError, match=msg):
-        s.map(lambda x: x, na_action="____")
 
 
 def test_apply_to_timedelta():
